+++ src/resources/views/shipping-method/_form.blade.php
@@ -51,6 +51,22 @@
 <hr>
 
 <div class="form-group row">
+    <label class="col-form-label col-form-label-sm col-md-2">{{ __('Restricted to Zone') }}</label>
+    <div class="col-md-10">
+        {{ Form::select('zone_id', $zones->pluck('name', 'id'), null, [
+                'class' => 'form-control form-control-sm' . ($errors->has('zone_id') ? ' is-invalid': ''),
+                'placeholder' => '<' . __('Unrestricted') . '>',
+           ])
+        }}
+        @if ($errors->has('zone_id'))
+            <div class="invalid-feedback">{{ $errors->first('zone_id') }}</div>
+        @endif
+    </div>
+</div>
+
+<hr>
+
+<div class="form-group row">
     <label class="col-form-label col-form-label-sm col-md-2">{{ __('Configuration') }}</label>
     <div class="col-md-10">
         <textarea name="configuration"